+++ vitest.config.ts
@@ -2,6 +2,7 @@ import { defineConfig } from 'vitest/config';
 import { fileURLToPath, URL } from 'node:url';
 
 export default defineConfig({
+  base: '/ganjoorSearch/',
   resolve: {
     alias: {
       '@': fileURLToPath(new URL('./src', import.meta.url)),